+++ test/execFilePromise.js
@@ -1,6 +1,0 @@
-'use strict';
-
-const { execFile } = require('child_process');
-const { promisify } = require('util');
-
-module.exports = promisify(execFile);