+++ test/execFilePromise.mjs
@@ -1,0 +1,4 @@
+import { execFile } from 'child_process';
+import { promisify } from 'util';
+
+export default promisify(execFile);