+++ test/index.js
@@ -1,9 +1,0 @@
-'use strict';
-
-const { TestDirector } = require('test-director');
-
-const tests = new TestDirector();
-
-require('./public/withGraphQLReact.test')(tests);
-
-tests.run();