+++ changelog.md
@@ -8,6 +8,7 @@
 - Updated the [`graphql-react`](https://npm.im/graphql-react) peer dependency to `^14.0.0`.
 - Updated dependencies, some of which require newer Node.js versions than were previously supported.
 - Replaced the the `package.json` `exports` field public [subpath folder mapping](https://nodejs.org/api/packages.html#packages_subpath_folder_mappings) (deprecated by Node.js) with a [subpath pattern](https://nodejs.org/api/packages.html#packages_subpath_patterns). Deep `require` paths within `next-graphql-react/public/` must now include the `.js` file extension.
+- The tests are now ESM in `.mjs` files instead of CJS in `.js` files.
 
 ### Patch